Add support for COG, and Degree, Radians
