+++ src/lib.rs
@@ -1,4 +1,6 @@
 #![warn(clippy::all, rust_2018_idioms)]
 
 mod app;
+mod communication;
+
 pub use app::TemplateApp;
